Création d'une bibliothèque clique
La bibliothèque calcule une clique max dans un graphe, implémentée de deux manières
- avec un algo exact recursif (cliquerecursif.c)
- avev un programme linéaire (cliquescip.c)

Le Makefile contient maintenant 2 sorties différentes selon qu'on souhaite la version classique
ou la version avec programme linéaire.
