+++ fonctions_mces.h
@@ -1,5 +1,5 @@
 #include "lecture_molecule_sdf.h"
-#include "graph.h"
+#include "helpers/graph.h"
 
 
 struct couple{
@@ -15,7 +15,7 @@ struct couple *construction_couples(struct molecule *M,int pos1, int pos2,int ta
 graph graphe_produit(int g1_chebi,int g2_chebi,struct molecule *M);
 void calcul_cl(graph m,int *dans_clique,int taille_clique,int *candidat,int taille_candidat, double date);	
 void la_clique_max(graph m,double date);
-int* graphe_g12(graph g12, struct molecule *M, int g1_chebi, int g2_chebi);
+int* graphe_g12(graph g12, int* clique_max, struct molecule *M, int g1_chebi, int g2_chebi);
 void liberer_molecule(struct molecule g);
 struct molecule * lecture_fichier_chebi();
 float mesure_similarite (int g1_chebi,int g2_chebi,struct molecule *M,double date,int taille_limite);
