+++ helpers/cliquescip.c
@@ -1,78 +1,10 @@
-#include "helpers/proglin_helper.h"
+#include "proglin_helper.h"
+#include "clique.h"
 #include "graph.h"
 #include <stdlib.h>
 #include <stdio.h>
-/**
-* Type concret représentant un graphe
-*/
 
-struct graph_s{
-  int n; /* Nombre de noeuds */
-  int** links; /* Matrice d'adjacence */
-};
-
-graph build_graph_from_matrix(int n, int** links){
-  graph g = (graph) malloc(sizeof(struct graph_s));
-  g->n = n;
-  
-  /* Attention à ne pas réécrire la matrice */
-  g->links = links;
-  
-  /* Version safe
-  int** linkscp = (int**) malloc(sizeof(int*) * n);
-  int i,j;
-  for(i = 0; i < n; i++){
-    linkscp[i] = (int*) malloc(sizeof(int) * n);
-    for(j = 0; j < n; j++){
-      linkscp[i][j] = links[i][j];
-    }
-  }
-  g->links = linkscp;
-  */
-  return g;
-}
-
-int nbnodes(graph g){
-  return g->n;
-}
-
-int** build_matrix_from_graph(graph g){
-
-  return g->links;
-   
-  /* Version safe
-  int n = nbnodes(g);
- 
-  int** linkscp = (int**) malloc(sizeof(int*) * n);
-  int i,j;
-  for(i = 0; i < n; i++){
-    linkscp[i] = (int*) malloc(sizeof(int) * n);
-    for(j = 0; j < n; j++){
-      linkscp[i][j] = g->links[i][j];
-    }
-  }
-  return linkscp;
-  */
-}
-
-void destroy(graph g){
-
-  /* Si links a été créé lors de la création du graphe
-  int n = nbnodes(g);
-  int i;
-  for(i = 0; i < n; i++){
-    free(g->links[i]);
-  } 
-  free(g->links);
-  */
-  free(g);
-}
-
-int are_neighbors(graph g, int u, int v){
-  return g->links[u][v];
-}
-
-int* maxclique(graph g, long timeout){
+int* clique_max(graph g, long timeout){
   /* Clique calculée avec un programme linéaire */
 
   
@@ -125,7 +57,8 @@ int* maxclique(graph g, long timeout){
 
   set_silent_verbose(pl);
   /* on indique le timeout en seconde */
-  set_max_timeout(pl, timeout);
+  if(timeout != 0)
+    set_max_timeout(pl, timeout);
 
   /* Résolution, ret est le code de retour, qui explique comment s'est terminé la résolution */
